Bug #17798: add support for 'o', 'g' and 'as' in dictionaries
This teaches the parameterised hash table about how to hash, compare and
free object paths and signatures, allowing them to be used as hash keys
and values, and also how to free strvs, so they can be used as values.
Adds some simple test methods which echo a{gas} and a{oas} dictionaries
back immediately and compare the results.