Bug #17795: depend on Glib 2.10 for hash unref consistency
Commit a8bf32 (for bug #11396) introduced a compile-time version check
for Glib 2.10 which changed whether hash_simple_free called
g_hash_table_unref or g_hash_table_destroy. This is at best useless for
people who actually wanted to rely on _unref being used instead of
_destroy, because dbus-glib can legitimately be built against older
versions of Glib than that in use by the application itself, and at
worst harmful as applications relying on the _unref semantics would have
their assumptions violated on platforms where this was the case. Given
in /most/ cases now, Glib 2.10 is readily available, and we were in this
kinda vague state where people could've used _unref and gotten away with
it, I'm going to say thats what our ABI is now, and stick to it. People
can depend on the next version of dbus-glib if they rely on _unref.