Bug 14579: remove pending call from hash table before cancelling it
Previously, the code implicitly assumed that cancelling the pending call
would not cause the DBusGProxy to be freed. This can fail if user_data
for the pending call holds the last reference to the DBusGProxy - in
this case, it is unsafe to be manipulating the contents of "priv" after
the call is cancelled.