Add GMainContext to dbus_g_bus_get_private, add a test case
To even sort of work with threads right now, a common workaround
is to open a private connection.  This patch more explicitly
supports creating a private connection, associating it with
the GMainContext which will be used for a thread.

Also, add a (very simple) test case which just uses a private
connection for the default main context.