core: allow duplicate property names on GInterfaces
Allows dbus-glib clients to handle objects that implement
more than one interface with the properties of the same name.
Normally this would not be allowed since with GObject all
properties are in the same namespace.  This patch allows
the interface to register "shadow" properties on a per-interface
basis which redirect the lookup of the property name.