Fix switching a connection's GMainContext
dbus_connection_setup_with_g_main and dbus_server_setup_with_g_main are
intended to switch the setup to a new main loop context if they are
called with a context other than the one with which they were initially
called.  However, this does not work in practice.  The problem has to do
with libdbus maintaining a list of watches associated with the
connection.  When dbus_connection_set_watch_functions is called, it
first adds the existing watches using the new function and data, then
removes the watches using the old function and data.  Remove_watch calls
connection_setup_remove_watch, and the latter does not check that the
watch's connection setup matches the connection setup passed to it, so
it winds up removing the watch from the new connection setup.  (The
watch has already been removed from the old setup at this point, as a
side effect of dbus_watch_set_data).

Fixes: <https://bugs.freedesktop.org/show_bug.cgi?id=30574>
Signed-off-by: Will Thompson <will.thompson@collabora.co.uk>