Run run-with-tmp-session-bus.sh directly rather than via ${SHELL}
It has a correct #!/bin/sh line and is executable, so just running it is
fine. This avoids running it under a user's chosen interactive shell,
which might be non-POSIX (zsh, csh). Gentoo Portage apparently sets
SHELL to /bin/false, presumably to trap this sort of thing.

(I've tested this on Debian unstable with /bin/sh -> /bin/dash, to check
that weren't accidentally relying on $SHELL being bash or similar.)

Based on a workaround from Myckel Habets on the Gentoo bug.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=27193
Bug-Gentoo: https://bugs.gentoo.org/show_bug.cgi?id=254192
Reviewed-by: Colin Walters <walters@verbum.org>