invoke_object_method: if marshalling an out argument fails, discard message
We shouldn't report this as OOM, because it probably wasn't (it's much
more likely to be programming error); we should also continue through the
arguments, so that we don't leak them.

By abandoning the message as soon as we detect a programming error,
we can use reply == NULL as an indicator of whether to keep appending.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=35767
Reviewed-by: Cosimo Alfarano <cosimo.alfarano@collabora.co.uk>