registrations test: fix race condition in adding match rules
If you don't block waiting for the reply to AddMatch, and you're using two
connections to the bus, then messages sent from connection 2 might not
be matched by connection 1, because they arrive before the AddMatch call
from connection 1. This manifests itself as the /registrations/twice
test *sometimes* failing (dependent on timing).

Only tests with two parallel bus connections need this, but to be safe,
I've added it throughout this test.

Reviewed-by: Vivek Dasmohapatra <vivek@collabora.co.uk>