test-variant-recursion: free our GValues and the output string correctly
Previously, if we'd freed the GValues, it would have crashed, because
g_value_take_boxed on a value of type G_TYPE_VALUE requires that the inner
GValue was g_malloc'd individually, but we were allocating them as a
block. This only "worked" because *none* of them were freed...

The returned string was also leaked; free it too, but only on success.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=41129
Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>