MyObject: make ThrowError, AsyncThrowError throw a caller-specified error
This can be used to test arbitrary errors, but only in-process; in
tests with the service out-of-process, like test-dbus-glib, the initial
error (matching the error they previously threw) will always be used.

This obsoletes ThrowErrorUnderscore, ThrowErrorMultiWord and
ThrowNotSupported, but not ThrowUnregisteredError due to some strange
assumptions about the validity of GError domains in that method
(see GNOME#660731).

Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>
Bug: https://bugs.freedesktop.org/show_bug.cgi?id=40151
Reviewed-by: Will Thompson <will.thompson@collabora.co.uk>