Fix two crashes when dbus_g_proxy_new_for_peer is used on a bus
The first part of the bug is that when NameOwnerChanged is received with
a dbus_g_proxy_new_for_peer (which has no name) alive, checking
whether it was affected by the NameOwnerChanged caused a NULL
dereference and segfault.

The second part of the bug is that if the last proxy in existence is
for a peer, when it was unregistered there would be no owner_match_rules,
causing a crash.

Both are exercised in the new test added here.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=41126
Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>
Reviewed-by: Will Thompson <will.thompson@collabora.co.uk>