Remove erroneous assert from connection_setup_add_timeout
When switching a connection from one main loop to another, timeouts are
moved to the new connection, so connection_setup_add_timeout gets called on
the new connection for all timeouts that were set on the old connection.
This means that the timeout's data can, in fact, be non-NULL, and it will be
freed / removed from the old connection as a side-effect of adding it to the
new connection. A similar assert was removed from dbus_connection_add_watch
a while ago as part of the original patch for this bug.

Also, clean up the regression test a bit, and test that a DBusConnection
can be migrated when a pending call is present.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=30574
[resolve conflicts -smcv]
Reviewed-by: Simon McVittie <simon.mcvittie@collabora.co.uk>