Avoid underlinking internal library libdbus-gtool.la
Anything that links libdbus-gtool needs to pull in GLib/GObject,
and the order matters (things that libdbus-gtool depends on must
appear after libdbus-gtool itself). libtool understands this, but
only if you tell it the full dependencies.

This broke compilation of test-dbus-glib.exe when cross-compiling
for 32-bit Windows with mingw-w64 (GNU tuple: i686-w64-mingw32).

The linking order used here for dependencies is "lowest in the stack
first", as recommended by the GStreamer documentation.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=68601
Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>