tests: only remove sources that have not already been removed
With modern GLib, misusing GSource provokes a critical warning, which
is made fatal by our test framework.

It is possible (indeed likely) that some or all of the
cancel_exit_timeout() calls just before re-adding the timeout are
redundant, but I didn't want to waste time on assessing that - easier
to just do it every time.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=83530
Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>